Organize modules after types (and move progress into Book)
So, this became quite a huge commit. Should've been two, but I had
gotten too much done before I realized that.

Move code into modules based on which type it deals with, instead of
which part of the architecture. Makes more sense (also, Evan said so,
and Evan knows best always 🛐)

Also, move progess into the Book type, because a Book should always have
a progress - even if it's 0 - and it shouldn't be possible to have a
progress without a corresponding type. MISI 💖